test_project.py
import pytest
from project import Account, deposit, withdraw, update, stocks

def test_account_number():
    account = Account()
    assert isinstance(account.number, int)
    assert len(str(account.number)) == 10  # Assuming account numbers are 10 digits.

def test_deposit():
    account = Account()
    deposit(account, 100)
    assert account.amount == 100

def test_withdraw():
    account = Account()
    deposit(account, 200)
    withdraw(account, 100)
    assert account.amount == 100

def test_withdraw_insufficient_funds():
    account = Account()
    deposit(account, 50)
    withdraw(account, 100)  # Attempt to withdraw more than the balance
    assert account.amount == 50  # Balance should remain the same

